+++ src/app/(content-layout)/[groupId]/tasklist/_tasklist/components/manage-task-item-modal/MangeTaskItemModal.tsx
@@ -6,7 +6,7 @@ import {
   ModalPortal,
   ModalTrigger,
 } from '@/components/common/modal';
-import ManageTaskItem from '@/components/manage-task-item/ManageTaskItem';
+import ManageTaskItem from '@/components/manage-task-item/components/ManageTaskItem';
 import useModalContext from '@/components/common/modal/core/useModalContext';
 import { Task } from '../../types/task-list-page-type';
 import Plus from '@/assets/Plus';
