+++ src/components/manage-task-item/components/TimePicker.tsx
@@ -1,6 +1,6 @@
 import clsx from 'clsx';
-import generateTime from './time-table';
-import { Time } from './type';
+import generateTime from '../time-table';
+import { Time } from '../type';
 
 interface TimePickerProps {
   selectedTime: Time;
